Fixed LCD test and added button
The display output for the LCD was formatting the temp and humidity
wrong. I've decided to skip using a string formatter and just print it
out in sections. I've also added in a basic button test that will use a
counter. Getting the temp and humidity take 250ms each, so a half a
second of our time is used getting sensor readings.

Buttons will have to be pressed for a continuous half-second in order to
be registered. A basic arduino will have almost exactly the number of
output pins to handle just temperature control. If we want to have
humidity control, something to increase the number of pins needs to be
added.