Moved pins and fixed bugs
Moved some pins around to avoid using pin 0, 1, and 13. Ended up having
to remove the humidity controls to get back those pins for now.

Made changes to how the cool down for the freezer works to use unsigned
longs. Before I had an int which made it not work correctly. Added in a
first_use variable so that the coil delay was ignored the first time the
program is run (when it's turned on).

Also made changes to deal with millis() rollover. This version is the
version that will go into the initial run of the cheese cave.