Added some stats and changed target temp band
We now display the duration of the last freezer run in seconds when the
freezer is not on. This will let me know how long on average the freezer
is on for.

Also in order to not get so cold, I've tightened the band from 2.5
seconds to 2 seconds. This should give us a range from 53F to 57F.